requirements.txt
# Data Quality Auto-Healing System - Requirements
# Python 3.10+ required

# Core ML/DL
torch==2.4.0
scikit-learn==1.5.0
numpy==1.26.4
pandas>=1.3.0,<2.2

# Big Data Processing
pyspark==3.5.3
pyarrow==17.0.0

# Streaming
kafka-python==2.0.2
confluent-kafka==2.6.0

# Orchestration
apache-airflow==2.10.0
apache-airflow-providers-apache-spark==4.10.0
apache-airflow-providers-openlineage>=1.8.0

# Database
psycopg2-binary==2.9.10
sqlalchemy>=1.4.36,<2.0
redis==5.2.0

# Monitoring
prometheus-client==0.21.0

# Data Validation
great-expectations==1.2.3

# MLOps
mlflow==2.18.0

# API
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
click==8.1.7

# Data Generation (for synthetic bad data)
faker==33.0.0
scipy==1.14.1

# Visualization (for notebooks)
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.24.0

# Testing
pytest==8.3.0
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.0
ipython==8.29.0
jupyter==1.1.1
notebook==7.3.0